requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-multipart==0.0.12
langchain-core==1.2.26
langchain-openai==1.1.12
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-text-splitters==0.3.11
sentence-transformers==3.3.1
faiss-cpu==1.13.2
pypdf==5.0.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
pytest==8.3.3
httpx==0.28.1
PyJWT==2.9.0
PyMySQL==1.1.1
